fix: make controls smaller
